Fixing unit test
