+++ src/sg/gov/acra/bizfiles/certificates/1.0/certificates-open-attestation.test.ts
@@ -1,6 +1,6 @@
 import schema from "./certificates-open-attestation.json";
 import sampleDocJson from "./certificates-open-attestation-sample-document.json";
-import { testOaCommon } from "../../common/tests/open-attestation.test";
+import { testOaCommon } from "../../common/tests/open-attestation";
 
 describe("common OA Tests", () => {
   testOaCommon(schema, sampleDocJson);
