+++ src/sg/gov/acra/bizfiles/common/tests/profile-schema.ts
@@ -10,6 +10,7 @@ function loadSchema(uri: string) {
 const ajv = new Ajv({ allErrors: true, loadSchema: loadSchema });
 let validator: Ajv.ValidateFunction;
 
+// eslint-disable-next-line jest/no-export
 export function testCommon(schema: object, sampleDocJson: object) {
   describe("commonTests", () => {
     beforeAll(async () => {
@@ -51,17 +52,65 @@ export function testCommon(schema: object, sampleDocJson: object) {
       });
     });
 
-    describe("entityType", () => {
-      it("should return array of errors without entityType", () => {
-        const badDoc = omit(cloneDeep(sampleDocJson), "entityType");
+    describe("status", () => {
+      it("should return array of errors without status", () => {
+        const badDoc = omit(cloneDeep(sampleDocJson), "status");
         expect(validator(badDoc)).toBe(false);
         expect(validator.errors).toStrictEqual([
           {
             keyword: "required",
             dataPath: "",
             schemaPath: "#/required",
-            params: { missingProperty: "entityType" },
-            message: "should have required property 'entityType'"
+            params: { missingProperty: "status" },
+            message: "should have required property 'status'"
+          }
+        ]);
+      });
+    });
+
+    describe("statusDate", () => {
+      it("should return array of errors without statusDate", () => {
+        const badDoc = omit(cloneDeep(sampleDocJson), "statusDate");
+        expect(validator(badDoc)).toBe(false);
+        expect(validator.errors).toStrictEqual([
+          {
+            keyword: "required",
+            dataPath: "",
+            schemaPath: "#/required",
+            params: { missingProperty: "statusDate" },
+            message: "should have required property 'statusDate'"
+          }
+        ]);
+      });
+    });
+
+    describe("address", () => {
+      it("should return array of errors without address", () => {
+        const badDoc = omit(cloneDeep(sampleDocJson), "address");
+        expect(validator(badDoc)).toBe(false);
+        expect(validator.errors).toStrictEqual([
+          {
+            keyword: "required",
+            dataPath: "",
+            schemaPath: "#/required",
+            params: { missingProperty: "address" },
+            message: "should have required property 'address'"
+          }
+        ]);
+      });
+    });
+
+    describe("activities", () => {
+      it("should return array of errors without activities", () => {
+        const badDoc = omit(cloneDeep(sampleDocJson), "activities");
+        expect(validator(badDoc)).toBe(false);
+        expect(validator.errors).toStrictEqual([
+          {
+            keyword: "required",
+            dataPath: "",
+            schemaPath: "#/required",
+            params: { missingProperty: "activities" },
+            message: "should have required property 'activities'"
           }
         ]);
       });
