+++ src/sg/gov/acra/bizfiles/profiles/business/1.0/business-profile-open-attestation.test.ts
@@ -1,6 +1,6 @@
 import schema from "./business-profile-open-attestation.json";
 import sampleDocJson from "./business-profile-open-attestation-sample-document.json";
-import { testOaCommon } from "../../../common/tests/open-attestation.test";
+import { testOaCommon } from "../../../common/tests/open-attestation";
 
 describe("common OA Tests", () => {
   testOaCommon(schema, sampleDocJson);
