+++ src/sg/gov/acra/bizfiles/profiles/company/1.0/company-profile-open-attestation.test.ts
@@ -1,6 +1,6 @@
 import schema from "./company-profile-open-attestation.json";
 import sampleDocJson from "./company-profile-open-attestation-sample-document.json";
-import { testOaCommon } from "../../../common/tests/open-attestation.test";
+import { testOaCommon } from "../../../common/tests/open-attestation";
 
 describe("common OA Tests", () => {
   testOaCommon(schema, sampleDocJson);
