+++ src/sg/gov/acra/bizfiles/profiles/company/1.0/sample-document.json
@@ -324,7 +324,7 @@
     }
   ],
   "receiptNumber": "ACRAXXXXXXXXXX06",
-  "receiptDate": "22/10/2019",
+  "receiptDate": "2019-10-22",
   "signatureName": "Tan Yong Tat",
   "qrCode": "asd",
   "authenticationNumber": "IXXXXXXX49"
