+++ src/sg/gov/acra/bizfiles/profiles/limited-liability-partnership/1.0/limited-liability-partnership-profile-open-attestation.test.ts
@@ -1,6 +1,6 @@
 import schema from "./limited-liability-partnership-profile-open-attestation.json";
 import sampleDocJson from "./limited-liability-partnership-open-attestation-sample-document.json";
-import { testOaCommon } from "../../../common/tests/open-attestation.test";
+import { testOaCommon } from "../../../common/tests/open-attestation";
 
 describe("common OA Tests", () => {
   testOaCommon(schema, sampleDocJson);
