+++ src/sg/gov/acra/bizfiles/profiles/limited-partnership/1.0/limited-partnership-profile-open-attestation.test.ts
@@ -1,6 +1,6 @@
 import schema from "./limited-partnership-profile-open-attestation.json";
 import sampleDocJson from "./limited-partnership-open-attestation-sample-document.json";
-import { testOaCommon } from "../../../common/tests/open-attestation.test";
+import { testOaCommon } from "../../../common/tests/open-attestation";
 
 describe("common OA Tests", () => {
   testOaCommon(schema, sampleDocJson);
