+++ src/sg/gov/acra/bizfiles/profiles/public-accountant-firm/1.0/public-accountant-firm-profile-open-attestation.test.ts
@@ -1,6 +1,6 @@
 import schema from "./public-accountant-firm-profile-open-attestation.json";
 import sampleDocJson from "./public-accountant-firm-open-attestation-sample-document.json";
-import { testOaCommon } from "../../../common/tests/open-attestation.test";
+import { testOaCommon } from "../../../common/tests/open-attestation";
 
 describe("common OA Tests", () => {
   testOaCommon(schema, sampleDocJson);
